chore(app):implement feedback and refactor for challenge 3 [Finishes #166832072]
