--- src/models/entities.js
@@ -1,34 +1,0 @@
-/* eslint-disable linebreak-style */
-const Users = [{
-  id: 1,
-  email: 'john@test.com',
-  firstName: 'john',
-  lastName: 'doe',
-  password: 'test',
-  address: '123 Testing Avenue, State, Country',
-  isAdmin: false,
-},
-{
-  id: 2,
-  firstName: 'Madike',
-  lastName: 'Johnson',
-  email: 'test@tester.com',
-  password: '$2b$10$5wgB0m73ao9vsTcypaLd3uvWOyZh0bRqWksnbT2irWH61KDrfPQHC',
-  address: 'Does it matter lol',
-  isAdmin: false,
-},
-{
-  id: 3,
-  firstName: 'James',
-  lastName: 'Potter',
-  email: 'jp@hogwarts.com',
-  password: '$2b$10$yK5w7gbn8jzgy7QSR4lJU.c6p2ixMooFOJMvK0hyIBAkHRzaElCgS',
-  address: 'Gryffindor, HOgwarts',
-  isAdmin: false,
-}];
-
-const entities = {
-  Users,
-};
-
-export default entities;
